Implement registration and authorization
+--------------------------------------------------------------------+
This commit implements registration and authorization for users.
To register, the user must specify an email name and password
in the request body, then a new user will be created in the database.
For authorization, you need to specify an email and password,
then the user receives a token for authorization.
+--------------------------------------------------------------------+
Endpoints:
+--------------------------------------------------------------------+

POST: /api/auth/sign_up
{
    name: "",
    "email": "",
    "password": ""
}

+--------------------------------------------------------------------+

POST: /api/auth/sign_in
{
    "email": "",
    "password": ""
}

+--------------------------------------------------------------------+